moved job info form to own folder and created component for next page
